Change the way 'scripts_version' is solved to allow minor versions to be found.
Previously only major versions were taken into consideration, like Harmony 20 would become 2000, Harmony 21 would be 2100, where the last two digits were hardcoded. But Harmony uses those digits for the minor versions, so Harmony 21.1 would be 2110 and with the previous implementation, this number was solved as 2100. This caused this version to be unable to read the shotgrid scripts in the right folder, thus, there was no shotgrid button.